Refactoring and removing old NEST
