+++ nest/__init__.py
@@ -1,1 +1,4 @@
-from nest.nest import NEST
+from nest.core.nest import NEST
+from nest.models import *
+
+__version__ = "0.1.0"
