+++ nest/core/__init__.py
@@ -1,0 +1,3 @@
+# Make important classes and functions available at nest.core
+from nest.core.nest import NEST
+from nest.core.exceptions import NESTError, ModelNotFoundError, ModelLoadError
