+++ nest/interfaces/__init__.py
@@ -1,0 +1,1 @@
+from nest.interfaces.base_model import BaseModelInterface
