+++ nest/models/__init__.py
@@ -1,0 +1,3 @@
+# This imports will trigger model registration
+import nest.models.fmri
+import nest.models.eeg
