+++ nest/models/eeg/__init__.py
@@ -1,0 +1,1 @@
+import nest.models.eeg.things_eeg
