+++ nest/models/fmri/__init__.py
@@ -1,0 +1,2 @@
+# Import models to register them
+import nest.models.fmri.nsd_fwrf
