+++ nest/models/fmri/fwrf/README.txt
