+++ nest/models/fmri/fwrf/config.py
