+++ nest/models/fmri/fwrf/file_utility.py
