+++ nest/models/fmri/fwrf/numpy_utility.py
