+++ nest/models/fmri/fwrf/plots.py
