+++ nest/models/fmri/fwrf/rf_grid.py
