+++ nest/models/fmri/fwrf/roi.py
