+++ nest/models/fmri/fwrf/torch_joint_training_unpacked_sequences.py
@@ -8,7 +8,7 @@ from tqdm import tqdm
 import math
 
 import numpy as np
-import nest.models.fwrf.numpy_utility as pnu
+import nest.models.fmri.fwrf.numpy_utility as pnu
 
 import torch as T
 import torch.nn as L
