+++ nest/models/model_cards/eeg_things_eeg_2_vit_b_32.yaml
