--- new_nest/__init__.py
@@ -1,4 +1,0 @@
-from new_nest.core.nest import NEST
-from new_nest.models import *
-
-__version__ = "0.1.0"
