--- new_nest/core/__init__.py
@@ -1,3 +1,0 @@
-# Make important classes and functions available at nest.core
-from new_nest.core.nest import NEST
-from new_nest.core.exceptions import NESTError, ModelNotFoundError, ModelLoadError
