--- new_nest/models/__init__.py
@@ -1,3 +1,0 @@
-# This imports will trigger model registration
-import new_nest.models.fmri
-import new_nest.models.eeg
