--- new_nest/models/eeg/__init__.py
@@ -1,1 +1,0 @@
-import new_nest.models.eeg.things_eeg
