--- new_nest/models/fmri/__init__.py
@@ -1,2 +1,0 @@
-# Import models to register them
-import new_nest.models.fmri.nsd_fwrf
